[PSL-755] [PSL-763] [PSL-764] (#177)
* [PSL-755] pasteld: adjust node deprecation policy for mainnet release

- set cNode deprecation time for mainnet to 2yrs from approx release height 500000
[PSL-763] Implement support for finding all tickets belonging to a specific collection in the find API
- implemented new RPC APIs:
  - tickets find nft "collection_activation_ticket_txid":
    returns all NFT registration tickets belonging to the collection specified by "collection_activation_ticket_txid".
    This API returns only NFT registration tickets - it does not check whether there are corresponding NFT activation tickets.
  - tickets find act "collection_activation_ticket_txid":
    returns all NFT activation tickets belonging to the collection specified by "collection_activation_ticket_txid".
  - tickets find action "collection_activation_ticket_txid":
    returns all Action Registration tickets (Sense) belonging to the collection specified by "collection_activation_ticket_txid".
	This API returns only Action registration tickets - it does not check whether corresponding Action activation tickets exist.
  - tickets find action-act "collection_activation_ticket_txid":
    returns all Action Activation tickets (Sense) belonging to the collection specified by "collection_activation_ticket_txid".
  - tickets find collection "collection_name"
    Find collection tickets by name - search is case-insensitive.
[PSL-764] Enhance 'tickets get' to return additional collection ticket information
    - RPC API [tickets get "collection_activate_ticket_txid"] now returns additional fields:
	   "activated_item_count": NNN, // number of activated items in this collection
	   "collection_state": "in-process|finalized|not-defined", // collection state
	   "is_expired_by_height": True|False, // True if collection is in finalized state because final allowed block height for the collection has been reached
	   "is_full": True|False, // True if collection is in finalized state because number of activated items in the collection reached max_collection_entries
	- max number of items in a collection is calculated now by activated items only, there can be more registered items in a collection than activated items
	- added nft|action collection item pre-registration check for max_collection_entries (calculated using activate tickets)
	  If collection item is registered, activation ticket for this item may not be validated if number of activated items in that collection reached max.
Other changes:
   - fixed typos (etherium -> ethereum)
   - fixed some compiler warnings
   - fixed an issue with "tickets list nft|act|action transferred", do not count tickets with item_copy_count=0
mn_tickets.py:
   - added tests for new "tickets find", test registers 3 items in the collection (nft|sense), but only activates 2, reaching max collection entries = 2.
   - added tests for tickets get "collection_activate_ticket_txid", check number of activated items, collection state by height and max entries

* mn_tickets fixes

* fix

* tickets_activate help for collections

+++ src/chain.cpp
@@ -36,7 +36,7 @@ void CChain::SetTip(CBlockIndex *pindex)
         vChain[pindex->nHeight] = pindex;
         pindex = pindex->pprev;
     }
-    gl_nChainHeight = vChain.size() - 1;
+    gl_nChainHeight = static_cast<uint32_t>(vChain.size()) - 1;
 }
 
 CBlockLocator CChain::GetLocator(const CBlockIndex *pindex) const {
